+++ EloquaClient/RestObject.cs
@@ -4,7 +4,7 @@ namespace Eloqua.Api.Rest.ClientLibrary
 {
     public class RestObject : IIdentifiable
     {
-        public int? id { get; set; }
+        public long? id { get; set; }
         public string name { get; set; }
 
         public string type;
